+++ wow_world_base/CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-* `AuctionHouse` type for Vanilla, TBC and Wrath.
+* `AuctionHouse` type for Vanilla, TBC and Wrath with member functions `deposit_percentage`, `cut_percentage`, and `(vanilla|tbc|wrath)_faction`.
 
 ## [0.1.1] - 2023-05-20